refactor(*): remove ktx imports
