+++ appdistribution/app/src/main/java/com/google/firebase/appdistributionquickstart/kotlin/KotlinMainActivity.kt
@@ -4,9 +4,9 @@ import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.appdistribution.FirebaseAppDistribution
 import com.google.firebase.appdistribution.FirebaseAppDistributionException
-import com.google.firebase.appdistribution.ktx.appDistribution
+import com.google.firebase.appdistribution.appDistribution
 import com.google.firebase.appdistributionquickstart.databinding.ActivityMainBinding
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.Firebase
 
 class KotlinMainActivity : AppCompatActivity() {
