+++ auth/app/src/main/java/com/google/firebase/quickstart/auth/kotlin/GenericIdpFragment.kt
@@ -26,9 +26,9 @@ import android.widget.ArrayAdapter
 import android.widget.Toast
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.auth.ktx.oAuthProvider
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.auth.oAuthProvider
+import com.google.firebase.Firebase
 import com.google.firebase.quickstart.auth.R
 import com.google.firebase.quickstart.auth.databinding.FragmentGenericIdpBinding
 import java.util.ArrayList
