+++ crash/app/src/main/java/com/google/samples/quickstart/crash/kotlin/MainActivity.kt
@@ -3,9 +3,9 @@ package com.google.samples.quickstart.crash.kotlin
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import com.google.firebase.crashlytics.FirebaseCrashlytics
-import com.google.firebase.crashlytics.ktx.crashlytics
-import com.google.firebase.crashlytics.ktx.setCustomKeys
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.crashlytics.crashlytics
+import com.google.firebase.crashlytics.setCustomKeys
+import com.google.firebase.Firebase
 import com.google.samples.quickstart.crash.databinding.ActivityMainBinding
 
 /**
