+++ database/app/src/main/java/com/google/firebase/quickstart/database/kotlin/BaseFragment.kt
@@ -3,8 +3,8 @@ package com.google.firebase.quickstart.database.kotlin
 import android.view.View
 import android.widget.ProgressBar
 import androidx.fragment.app.Fragment
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.auth.auth
+import com.google.firebase.Firebase
 
 open class BaseFragment : Fragment() {
     private var progressBar: ProgressBar? = null
