+++ firestore/app/src/main/java/com/google/firebase/example/fireeats/kotlin/adapter/RatingAdapter.kt
@@ -6,7 +6,7 @@ import androidx.recyclerview.widget.RecyclerView
 import com.google.firebase.example.fireeats.databinding.ItemRatingBinding
 import com.google.firebase.example.fireeats.kotlin.model.Rating
 import com.google.firebase.firestore.Query
-import com.google.firebase.firestore.ktx.toObject
+import com.google.firebase.firestore.toObject
 import java.text.SimpleDateFormat
 import java.util.Locale
