+++ inappmessaging/app/src/main/java/com/google/firebase/fiamquickstart/kotlin/KotlinMainActivity.kt
@@ -5,13 +5,13 @@ import android.util.Log
 import androidx.appcompat.app.AppCompatActivity
 import com.google.android.material.snackbar.Snackbar
 import com.google.firebase.analytics.FirebaseAnalytics
-import com.google.firebase.analytics.ktx.analytics
+import com.google.firebase.analytics.analytics
 import com.google.firebase.fiamquickstart.R
 import com.google.firebase.fiamquickstart.databinding.ActivityMainBinding
 import com.google.firebase.inappmessaging.FirebaseInAppMessaging
-import com.google.firebase.inappmessaging.ktx.inAppMessaging
-import com.google.firebase.installations.ktx.installations
-import com.google.firebase.ktx.Firebase
+import com.google.firebase.inappmessaging.inAppMessaging
+import com.google.firebase.installations.installations
+import com.google.firebase.Firebase
 
 class KotlinMainActivity : AppCompatActivity() {
